include more information about fetch on workers
I was quite confused by this example, since I didn't realize that there is a `getindex(f::Future, args...) = getindex(fetch(f), args...)`. 
Because of this I misinterpreted the example (my first conclusion: `remote_call` must magically unpack the `Future`). 
Also, it didn't really make clear, that fetch turns into a noop if applied to a `Future` from the same worker.
I tried to add this information in a concise way. 
If I failed at this, it'd be nice if someone else could rewrite this text ;)